start implement worker system

+++ yu/src/core/crc32.cpp
@@ -146,7 +146,7 @@ void CRC32_Update( unsigned long &crcvalue, const u8 data )
 	crcvalue = crctable[ ( crcvalue ^ data ) & 0xff ] ^ ( crcvalue >> 8 );
 }
 
-void CRC32_UpdateChecksum( unsigned long &crcvalue, const void *data, int length ) {
+void CRC32_UpdateChecksum(unsigned long &crcvalue, const void *data, size_t length) {
 	unsigned long crc;
 	const unsigned char *buf = (const unsigned char *) data;
 
@@ -161,7 +161,7 @@ void CRC32_FinishChecksum( unsigned long &crcvalue ) {
 	crcvalue ^= CRC32_XOR_VALUE;
 }
 
-unsigned long CRC32_BlockChecksum( const void *data, int length ) {
+unsigned long CRC32_BlockChecksum(const void *data, size_t length) {
 	unsigned long crc;
 
 	CRC32_InitChecksum( crc );
@@ -170,7 +170,7 @@ unsigned long CRC32_BlockChecksum( const void *data, int length ) {
 	return crc;
 }
 
-unsigned long CRC32( const void *data, int length )
+unsigned long CRC32(const void *data, size_t length)
 {
 	return CRC32_BlockChecksum(data, length);
 }
